Added placholder pages, moved default style to root src
